修正

+++ ros2_humble_ws/src/flashgo/src/flashgo_client.cpp
@@ -1,23 +1,22 @@
 /*
-* @file flashgo_client.cpp
-* @author huigang39
-* @brief ROS2 client for FlashGo
-* @date 2023-10-19
-* @version 0.0.1
-* @note ROS2 client for FlashGo
-* @warning None.
-* @bug None.
-*/
+ * @file flashgo_client.cpp
+ * @author huigang39
+ * @brief ROS2 client for FlashGo
+ * @date 2023-10-19
+ * @version 0.0.1
+ * @note ROS2 client for FlashGo
+ * @warning None.
+ * @bug None.
+ */
 
 #include "rclcpp/rclcpp.hpp"
 #include "sensor_msgs/msg/laser_scan.hpp"
-#include <cmath>
 
-#define RAD2DEG(x) ((x)*180./M_PI)
+#define RAD2DEG(x) ((x) * 180. / M_PI)
 
 int round_double(double number)
 {
-    return (number > 0.0) ? static_cast<int>(number + 0.5) : static_cast<int>(number - 0.5);
+    return (number > 0.0) ? (number + 0.5) : (number - 0.5);
 }
 
 void scanCallback(const sensor_msgs::msg::LaserScan::SharedPtr scan)
@@ -26,7 +25,8 @@ void scanCallback(const sensor_msgs::msg::LaserScan::SharedPtr scan)
     RCLCPP_INFO(rclcpp::get_logger("eai_info"), "I heard a laser scan %s[%d]:", scan->header.frame_id.c_str(), count);
     RCLCPP_INFO(rclcpp::get_logger("eai_info"), "angle_range, %f, %f", RAD2DEG(scan->angle_min), RAD2DEG(scan->angle_max));
 
-    for (int i = 0; i < count; i++) {
+    for (int i = 0; i < count; i++)
+    {
         float degree = RAD2DEG(scan->angle_min + scan->angle_increment * i);
         RCLCPP_INFO(rclcpp::get_logger("eai_info"), ": [%f, %f]", degree, scan->ranges[i]);
     }
